+++ octobot_trading/orders/order_adapter.pxd
@@ -15,11 +15,11 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-cpdef float adapt_price(dict symbol_market, float price)
-cpdef float adapt_quantity(dict symbol_market, float quantity)
-cpdef float trunc_with_n_decimal_digits(object value, int digits)
-cpdef list adapt_order_quantity_because_quantity(object limiting_value, object max_value, object quantity_to_adapt, float price, dict symbol_market)
-cpdef list adapt_order_quantity_because_price(object limiting_value, object max_value, float price, dict symbol_market)
-cpdef list split_orders(object total_order_price, object max_cost, object valid_quantity, object max_quantity, float price, object quantity, dict symbol_market)
-cpdef object add_dusts_to_quantity_if_necessary(object quantity, float price, dict symbol_market, object current_symbol_holding)
-cpdef check_and_adapt_order_details_if_necessary(object quantity, float price, dict symbol_market, object fixed_symbol_data=*)
+cpdef double adapt_price(dict symbol_market, double price)
+cpdef double adapt_quantity(dict symbol_market, double quantity)
+cpdef double trunc_with_n_decimal_digits(object value, int digits)
+cpdef list adapt_order_quantity_because_quantity(object limiting_value, object max_value, object quantity_to_adapt, double price, dict symbol_market)
+cpdef list adapt_order_quantity_because_price(object limiting_value, object max_value, double price, dict symbol_market)
+cpdef list split_orders(object total_order_price, object max_cost, object valid_quantity, object max_quantity, double price, object quantity, dict symbol_market)
+cpdef object add_dusts_to_quantity_if_necessary(object quantity, double price, dict symbol_market, object current_symbol_holding)
+cpdef check_and_adapt_order_details_if_necessary(object quantity, double price, dict symbol_market, object fixed_symbol_data=*)
